requirements.txt
fastapi
uvicorn
reportlab
python-multipart
aiofiles
jinja2
google-cloud-storage
weasyprint
pydantic
pytest